+++ R/data.R
@@ -33,6 +33,21 @@
 #' @keywords data
 "partnerships"
 
+#' Recid name / description lookup
+#'
+#' A [tibble][tibble::tibble-package] containing the names and description of the recids.
+#' To be used with the \code{recid} parameter when reading the episode files.
+#'
+#' @format A [tibble][tibble::tibble-package] with 18 rows and 2 variables:
+#' \describe{
+#'   \item{recid}{The short-form recid}
+#'   \item{Description}{A description of the recid}
+#' }
+#' @name recids
+#' @docType data
+#' @keywords data
+"recids"
+
 #' A tibble containing an example cohort of CHI numbers under different
 #' variable names. It is used for testing and to illustrate examples.
 #' It is likely that many of the 'chi numbers' are not valid.