+++ NEWS.md
@@ -1,3 +1,5 @@
+# slfhelper (development version)
+
 # slfhelper 0.6.0
 
 * Fix bug with selecting partnerships/recids when also selecting multiple years (#13) (#14)