+++ NEWS.md
@@ -1,6 +1,7 @@
 # slfhelper (development version)
 
 * Add check for server and hscdiip access on package load.
+* Update variable lists.
 
 # slfhelper 0.6.0
 
@@ -10,5 +11,5 @@
 
 * Added a `NEWS.md` file to track changes to the package.
 * Moved to use Rmd version for the readme.
-* Bump dplyr version requirement to 1.0.0 to enable use of `across()` etc.
+* Bump dplyr version requirement to 1.0.0 to enable use of `across()` and other new things.
 * Add checks to the year