+++ NEWS.md
@@ -3,6 +3,7 @@
 * Add check for server and hscdiip access on package load.
 * Update variable lists.
 * Add a recid lookup for use with the `recid` filtering parameter.
+* Now depends on R >= 3.5.1 
 
 # slfhelper 0.6.0